CURR-60 Spring Security authentication

+++ src/main/java/net/ukr/dreamsicle/repository/CurrencyRepository.java
@@ -3,5 +3,5 @@ package net.ukr.dreamsicle.repository;
 import net.ukr.dreamsicle.model.Currency;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CurrencyRepository extends JpaRepository<Currency, Integer> {
+public interface CurrencyRepository extends JpaRepository<Currency, Long> {
 }
